% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggca.R
\name{ggca}
\alias{ggca}
\title{\code{"CA"} ordination plot}
\usage{
ggca(ord, groups = NULL, axes = c(1, 2), scaling = 2, obslab = FALSE,
  moblabs = NULL, obssize = 2, obscol = "black", obspch = 16,
  obsFonts = "serif", obsface = "plain", spe = TRUE, msplabs = NULL,
  spearrow = 0.2, spelab = TRUE, spmapsize = NULL, spaline = 1,
  spalwd = 0.5, spacol = "grey30", sprotate = NULL, spesize = 4,
  specol = "red", spepch = 16, speFonts = "serif", speface = "plain",
  envs = NULL, mflabs = NULL, farrow = 0.2, fmapsize = NULL,
  faline = 1, falwd = 0.5, facol = "blue", fzoom = 1, frotate = NULL,
  fsize = 5, fcol = "blue", fFonts = "serif", fface = "plain",
  ellipse = FALSE, ellprob = 0.95, cirlwd = 1, cirline = 2)
}
\arguments{
\item{ord}{An object produced by \code{\link[vegan]{cca}}.}

\item{groups}{An grouping factor, its length is the same as the row number of the ordination dataframe.}

\item{axes}{Axes shown.}

\item{scaling}{Scaling for species and site scores. Either species (2) or site (1) scores are scaled by eigenvalues,
and the other set of scores is left unscaled, or with 3 both are scaled symmetrically by square root of eigenvalues.
Unscaled raw scores stored in the result can be accessed with scaling = 0.
The type of scores can also be specified as one of "none", "sites", "species", or "symmetric",
which correspond to the values 0, 1, 2, and 3 respectively.}

\item{obslab}{A logical value, obslab = FALSE(The row variables are displayed as points), obslab = TRUE(The row variables are displayed as texts).}

\item{moblabs}{A vector of strings, rename the row variable names displayed.}

\item{obssize}{The size of row variables.}

\item{obscol}{The colour of row variables.}

\item{obspch}{The point shape of row variables.}

\item{obsFonts}{The family of row variables.}

\item{obsface}{The fontface of row variables.}

\item{spe}{A logical value, whether the column variables are displayed.}

\item{msplabs}{A vector of strings, rename the col variable names displayed.}

\item{spearrow}{Arrowhead length of col variables.}

\item{spelab}{A logical value, spelab = FALSE(The col variables are displayed as points), spelab = TRUE(The col variables are displayed as texts).}

\item{spmapsize}{Numeric value, the size of col variable labels is mapped by the length of arrowhead.}

\item{spaline}{Type of arrowhead segment.}

\item{spalwd}{Numeric value, the width of arrowhead segment.}

\item{spacol}{The colour of arrowhead segment.}

\item{sprotate}{Numeric value, rotation angle of col variable labels.}

\item{spesize}{Numeric value, the size of col variable labels or points.}

\item{specol}{The colour of col variable labels or opoints.}

\item{spepch}{Type of col variable labels or points .}

\item{speFonts}{The family of col variable labels.}

\item{speface}{The fontface of col variable labels.}

\item{envs}{Dataframe fitted.}

\item{mflabs}{A vector of strings, rename the fitted variable names displayed.}

\item{farrow}{Arrowhead length of fitted variables.}

\item{fmapsize}{Numeric value, the size of fitted variable labels is mapped by the length of arrowhead.}

\item{faline}{Arrowhead type of fitted variables.}

\item{falwd}{Numeric value, arrowhead width of fitted variables.}

\item{facol}{Arrowhead colour of fitted variables.}

\item{fzoom}{Numeric value, scaling arrow length.}

\item{frotate}{Numeric value, rotation angle of fitted variable labels.}

\item{fsize}{Numeric value, the size of fitted variable labels or points.}

\item{fcol}{The colour of fitted variable labels or opoints.}

\item{fFonts}{The family of fitted variable labels.}

\item{fface}{The fontface of fitted variable labels.}

\item{ellipse}{A logical value, whether confidence ellipses are displayed.}

\item{ellprob}{Numeric value, confidence interval.}

\item{cirlwd}{Numeric value, line width of ellipse.}

\item{cirline}{Line type of ellipse.}
}
\value{
Returns a ggplot object.
}
\description{
Output ca ordination plot produced by \code{\link[vegan]{cca}}.
}
\details{
Control of some parts is put in the function. Theme of figure can be set by theme of ggplot2.
}
\examples{
 
data(Spes)
library(vegan)

#get group factor
Spe.w <- hclust(dist(scale(Spes)), "ward.D")
gr <- cutree(Spe.w , k=4)
grl <- factor(gr)

# Compute CCA
Spe.ca <- cca(Spes,scale = TRUE)
head(summary(Spe.ca))

# Produce a plot 
ggca(Spe.ca)

# Add a group
ggca(Spe.ca, group = grl) 

# Set a theme 
require(ggplot2)
ggca(Spe.ca, group = grl, fcol = "white", facol = "white") + theme_dark()

# Remove the arrow
ggca(Spe.ca, group = grl, spearrow = NULL)

# Modify legend title, group color and point shape
ggca(Spe.ca, group = grl, spearrow = NULL) + 
  scale_color_manual(name = "Groups",values = c("red2", "purple1", "grey20","cyan")) +
  scale_shape_manual(name = "Groups",values = c(8,15,16,17))

#Add confidence ellipses
ggca(Spe.ca, group = grl, spearrow = NULL, ellipse = TRUE) +
  scale_colour_hue(l = 70, c = 300)
}
\author{
Dongya Wang <wdy91617@163.com>
}
