% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpcoa.R
\name{ggpcoa}
\alias{ggpcoa}
\title{\code{"PCoA"} ordination plot}
\usage{
ggpcoa(ord, ordata, groups = NULL, axes = c(1, 2), obslab = FALSE,
  moblabs = NULL, obssize = 2, obscol = "black", obspch = 16,
  obsFonts = "serif", obsface = "plain", spe = TRUE, msplabs = NULL,
  spearrow = 0.2, spelab = TRUE, spmapsize = NULL, spaline = 1,
  spalwd = 0.5, spacol = "grey30", sprotate = NULL, spesize = 4,
  specol = "red", spepch = 16, speFonts = "serif", speface = "plain",
  envs = NULL, mflabs = NULL, farrow = 0.2, fmapsize = NULL,
  faline = 1, falwd = 0.5, facol = "blue", fzoom = 1, frotate = NULL,
  fsize = 5, fcol = "blue", fFonts = "serif", fface = "plain",
  ellipse = FALSE, ellprob = 0.95, cirlwd = 1, cirline = 2)
}
\arguments{
\item{ord}{An object produced by \code{\link[stats]{cmdscale}}.}

\item{ordata}{An dataframe that is provided to function \code{\link[vegan]{wascores}}.}

\item{groups}{An grouping factor, its length is the same as the row number of the ordination dataframe.}

\item{axes}{Axes shown.}

\item{obslab}{A logical value, obslab = FALSE(The row variables are displayed as points), obslab = TRUE(The row variables are displayed as texts).}

\item{moblabs}{A vector of strings, rename the row variable names displayed.}

\item{obssize}{The size of row variables.}

\item{obscol}{The colour of row variables.}

\item{obspch}{The point shape of row variables.}

\item{obsFonts}{The family of row variables.}

\item{obsface}{The fontface of row variables.}

\item{spe}{A logical value, whether the column variables are displayed.}

\item{msplabs}{A vector of strings, rename the col variable names displayed.}

\item{spearrow}{Arrowhead length of col variables.}

\item{spelab}{A logical value, spelab = FALSE(The col variables are displayed as points), spelab = TRUE(The col variables are displayed as texts).}

\item{spmapsize}{Numeric value, the size of col variable labels is mapped by the length of arrowhead.}

\item{spaline}{Type of arrowhead segment.}

\item{spalwd}{Numeric value, the width of arrowhead segment.}

\item{spacol}{The colour of arrowhead segment.}

\item{sprotate}{Numeric value, rotation angle of col variable labels.}

\item{spesize}{Numeric value, the size of col variable labels or points.}

\item{specol}{The colour of col variable labels or opoints.}

\item{spepch}{Type of col variable labels or points .}

\item{speFonts}{The family of col variable labels.}

\item{speface}{The fontface of col variable labels.}

\item{envs}{Dataframe fitted.}

\item{mflabs}{A vector of strings, rename the fitted variable names displayed.}

\item{farrow}{Arrowhead length of fitted variables.}

\item{fmapsize}{Numeric value, the size of fitted variable labels is mapped by the length of arrowhead.}

\item{faline}{Arrowhead type of fitted variables.}

\item{falwd}{Numeric value, arrowhead width of fitted variables.}

\item{facol}{Arrowhead colour of fitted variables.}

\item{fzoom}{Numeric value, scaling arrow length.}

\item{frotate}{Numeric value, rotation angle of fitted variable labels.}

\item{fsize}{Numeric value, the size of fitted variable labels or points.}

\item{fcol}{The colour of fitted variable labels or opoints.}

\item{fFonts}{The family of fitted variable labels.}

\item{fface}{The fontface of fitted variable labels.}

\item{ellipse}{A logical value, whether confidence ellipses are displayed.}

\item{ellprob}{Numeric value, confidence interval.}

\item{cirlwd}{Numeric value, line width of ellipse.}

\item{cirline}{Line type of ellipse.}
}
\value{
Returns a ggplot object.
}
\description{
Output pcoa ordination plot.
}
\details{
Control of some parts is put in the function. Theme of figure can be set by theme of ggplot2.
}
\examples{
data(Envs)
library(vegan)
# get group factor
Env.w <- hclust(dist(scale(Envs)), "ward.D")
gr <- cutree(Env.w , k=4)
grl <- factor(gr)

# Compute PCoA
Env.bray <- vegdist(Envs)
Env.pcoa <- cmdscale(Env.bray, eig=TRUE)
summary(Env.pcoa)

# Produce a plot 
ggpcoa(Env.pcoa, ordata = Envs)

# Add a group
ggpcoa(Env.pcoa, ordata = Envs, group = grl) 

# Set a theme 
require(ggplot2)
ggpcoa(Env.pcoa, ordata = Envs, group = grl, fcol = "white", facol = "white") + theme_dark()

# Remove the arrow
ggpcoa(Env.pcoa, ordata = Envs, group = grl, spearrow = NULL)

# Modify legend title, group color and point shape
ggpcoa(Env.pcoa, group = grl, ordata = Envs, spearrow = NULL) + 
  scale_color_manual(name = "Groups",values = c("red2", "purple1", "grey20","cyan")) +
  scale_shape_manual(name = "Groups",values = c(8,15,16,17))

#Add confidence ellipses
ggpcoa(Env.pcoa, ordata = Envs, group = grl, spearrow = NULL, ellipse = TRUE) +
  scale_colour_hue(l = 70, c = 300)
}
\author{
Dongya Wang <wdy91617@163.com>
}
